requirements.txt
requests~=2.31.0
PyYAML~=6.0.1
redis~=5.0.8
urllib3~=2.0.7